requirements.txt
daft
pylance
transformers
torch
torchvision
pillow
numpy
pyarrow